[JA] 21.04.13 HW UPDATE - 순서 바꿈
