+++ src/_210413/boj17143/Main_ja.java
@@ -3,108 +3,132 @@ package _210413.boj17143;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.PriorityQueue;
 
 /**
- * [골드4] 중량제한
+ * [골드2] 낚시왕
  * 1. 결과 : 성공
- * 2. 시간 복잡도 : O(ElogE) => (10^5)log(10^5) = 5*10^5...?
- * 	- 이유 
- *	  - 다익스트라로 시작 지점에서 끝지점까지의 최대 중량제한 구하기
- *	  - 다익스트라 : 우선 순위 큐 + 인접 리스트
- *		-> 우선 순위 큐의 추가 또는 삭제 O(logE) 
- *		-> E개의 간선에 대해 이 작업을 진행하므로 O(E), 전체 시간복잡도는 O(ElogE)
- *			=> 보통 E <= V^2 이므로, O(ElogV)로 볼 수 있다고 함.
- *		
- *		* 인접 행렬 사용 시, 메모리초과
- *			-> 다리 중복 정보 때문에 인접 행렬로 저장하려다가, 메모리 초과남
- *			-> 10,000 * 10,000 * 4 = 400,000,000 = 400MB (제한 : 128MB)
- * 	- 범위 
- *    - N : ~ 10,000
- *    - M : ~ 100,000
- *   
- *  - 접근 방식
- *    1) 섬 사이의 다리 정보를 인접 리스트로 저장 (중복도 저장)
- *    2) 우선순위큐에 시작지점 추가
- *    3) 다익스트라
- *    	 - 우선순위큐로, 누적 중량제한이 제일 큰 노드부터 pop
- *    	 - 거쳐 온 경로의 중량제한이 dist보다 큰 경우만 dist update
- *    	 - 도착지에 도착하면 중력 출력 후 return
+ * 2. 시간 복잡도 : C*(R*C)*(R*2-1) = (CR)^2 = 100,000,000 ...?
+ * 	  - 이유
+ * 		접근 방식을 토대로 나옴.
+ * 	
+ * 	  - 범위
+ * 		- R,C : ~ 100
+ * 		- M : ~ R*C
+ * 		- S : ~ 1,000
  * 
+ * 	  - 접근 방식
+ * 		- C만큼 반복(낚시왕 이동)
+ * 		  1) 해당 열의 최 상단 상어 제거
+ * 		  2) 상어 이동
+ * 		  -> R*C 전체 탐색
+ * 		  -> 상어 만나면 이동, 최대 R*2-2만큼 이동
+ * 	
+ * 	  - 최적화
+ * 		* 시간  
+ * 		  1) 속력(s) 그대로 한칸씩 이동 -> 2552ms
+ * 		  2) 속력(s)이 R*2-2인 경우, 처음 방향을 그대로 유지하면서 제자리로 돌아온다. (열 이동인 경우, C*2-2)
+ * 			 이러한 논리를 토대로, s%=(R*2-2)으로 이동 횟수를 줄임 -> 824ms
+ * 
+ * 		* 메모리
+ * 		  1) 이동해야 하는 상어를 별도의 ArrayList에 저장한 후, 
+ * 			  상어 한 마리씩 이동해서 기존 map에 적용 (상어를 clone까지 해야함) -> 80,704 KB
+ * 		  2) 새로운 map에다가 상어를 이동신 후,
+ * 			  새로운 map을 기존 map에다가 덮어쓰기 -> 35,680 KB
+ * 			
  * 3. 후기
- *  - 답은 맞게 나오는데 자꾸 틀려서 고생했다. 
- *    알고보니 "서로 같은 두 도시 사이에 여러 개의 다리가 있을 수도 있으며"가 중복 입력이 들어올 수 있다는 말이었다. :(
- *    몰라도, 통과할 수 있었는데 조건에서 부등호를 반대로 썼었다. 으휴
- *     
- *  - 그 후에는 계속 시초가 나서 고생했다. 무려 71%까지 통과했다! 왜 나는지 납득할 수 없었다. (그래서 다익으로 풀고 말겠다고 고집부림)
- *    보니까 반환 실수더라. 도착지점 만나면 return을 해야 하는데 break를 했다. 뭉충이 :< 
- *    
- *  - 자꾸 자잘한 실수를 한다. 꼼꼼하게 코드 보는 습관 좀 들여야겠다..
+ * - 아기 상어랑 비슷해서 구현 자체는 어렵지 않았다. (다행)
+ * - 시간이 좀 걸렸던 이유는, 속력을 좀 최적화하고 싶은 마음에 고집 좀 부렸다.
+ * 	 -> 최적화를 안하고, 속력 그대로 이동해도 통과는 한다! 그러나 아슬아슬하게 한다. 잘못하면  시초난다. 
+ * 
+ * - 배열 map 메모리 아끼겠다고 상어 복사하고 별 짓을 다했는데, 
+ *   오히려 새로운 map을 만드는게 메모리 절약되니까 약간 허무하면서 신기했다.
+ * 	 -> 상어 복사하면서 객체 복사 어떻게 하는걸까 찾아보니까, 
+ * 		Copyable 인터페이스의 copy 메서드를 구현해서 사용하드라! js는 쉽게 되는데 ^^;;
  */
 public class Main_ja {
 	public static int[] inputToIntArr(String str) {
 		return Arrays.stream(str.split(" ")).mapToInt(Integer::parseInt).toArray();
-	}
+	};
+
+	static int[][] dt = { {}, { -1, 0 }, { 1, 0 }, { 0, 1 }, { 0, -1 } };
+	static int[] reverseDir = { 0, 2, 1, 4, 3 };
+	static int R, C;
+
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int[] input = inputToIntArr(br.readLine());
-		int N = input[0];
-		int M = input[1];
-		
-		ArrayList<Node>[] bridge = new ArrayList[N+1];
-		for (int i = 0; i < N+1; i++) {
-			bridge[i] = new ArrayList<Node>();
-		}
-		// 인접 리스트
+		R = input[0];
+		C = input[1];
+		int M = input[2];
+
+		Shark[][] map = new Shark[R][C];
 		for (int i = 0; i < M; i++) {
 			input = inputToIntArr(br.readLine());
-			int A = input[0];
-			int B = input[1];
-			int W = input[2];
-			bridge[A].add(new Node(B,W));
-			bridge[B].add(new Node(A,W));
+			int r = input[0] - 1;
+			int c = input[1] - 1;
+			map[r][c] = new Shark(input[2], input[3], input[4]);
+			// 제자리로 돌아오는 속력은 제외해서 저장
+			map[r][c].s %= 2 * ((map[r][c].d < 3 ? R : C) - 1);
 		}
 		
-		input = inputToIntArr(br.readLine());
-		int start = input[0];
-		int end = input[1];
-		
-		PriorityQueue<Node> pq = new PriorityQueue<>();
-		boolean[] visited = new boolean[N+1];
-		int[] dist = new int[N+1];
-		Arrays.fill(dist, -1);
-		
-		// 시작지점 추가
-		pq.add(new Node(start, 1_000_000_001));
-		
-		// 다익스트라
-		while(!pq.isEmpty()) {
-			Node cur = pq.poll();
-			if(cur.v == end) { // 도착지점 도착 시, 출력 및 종료
-				System.out.println(cur.weight);
-				return;
-			}
-			for(Node next : bridge[cur.v]) {
-				int w = Math.min(cur.weight, next.weight); 
-				if(visited[next.v] || dist[next.v] >= w) continue; // 방문 또는 최대 중량 제한보다 작거나 같은 경우 pass
-				dist[next.v] = Math.max(w, dist[next.v]);
-				pq.add(new Node(next.v, w));
-			}
-			visited[cur.v] = true;
+		int result = 0;
+		for (int c = 0; c < C; c++) {
+			result += fishing(map, c); // 낚시
+			map = moveShark(map); // 상어 이동
 		}
+		System.out.println(result);
 	}
-	static class Node implements Comparable<Node>{
-		int v;
-		int weight;
-		public Node(int v, int weight) {
-			this.v = v;
-			this.weight = weight;
+
+	public static int fishing(Shark[][] map, int c) {
+		for (int r = 0; r < R; r++) { // 해당 열에서 젤 위의 상어 제거
+			Shark target = map[r][c];
+			if (target == null)
+				continue;
+			map[r][c] = null;
+			return target.z;
 		}
-		@Override
-		public int compareTo(Node o) {
-			return o.weight - this.weight;
+		return 0;
+	}
+
+	public static Shark[][] moveShark(Shark[][] map) {
+		Shark[][] newMap = new Shark[R][C]; // 상어가 이동할 새로운 맵
+		for (int i = 0; i < R; i++) {
+			for (int j = 0; j < C; j++) {
+				Shark target = map[i][j];
+				if (target == null) continue;
+				
+				int r = i;
+				int c = j;
+				int s = target.s;
+				while (s-- > 0) { // 속력만큼 상어 이동
+					int mr = r + dt[target.d][0];
+					int mc = c + dt[target.d][1];
+					if (mr < 0 || mc < 0 || mr >= R || mc >= C) { // 범위 넘어가면 방향 바꾸고 다시 이동
+						target.d = reverseDir[target.d];
+						s++;
+						continue;
+					}
+					r = mr;
+					c = mc;
+				}
+				if (newMap[r][c] != null && newMap[r][c].z > target.z) // 이동할 자리에 상어 있으면 크기 비교
+					continue;
+				newMap[r][c] = target;
+			}
+		}
+		return newMap;
+	}
+
+	static class Shark {
+		int s;
+		int d;
+		int z;
+
+		public Shark(int s, int d, int z) {
+			this.s = s;
+			this.d = d;
+			this.z = z;
 		}
 	}
 }
